feat: we should be using a `TypeError`
`TypeError` extends the `Error` object. But semantiaclly more accurate
in this case.

MDN describes this as:

"The TypeError object represents an error when an operation could not be
performed, typically (but not exclusively) when a value is not of the
expected type."

As the length of the array of values does not equal that of the request,
it is actually better to throw this `TypeError`.